+++ instrumentation/opentelemetry-instrumentation-urllib3/tests/test_urllib3_integration.py
@@ -19,12 +19,12 @@ import httpretty
 import urllib3
 import urllib3.exceptions
 
-from opentelemetry import context, trace
-
-# FIXME: fix the importing of this private attribute when the location of the _SUPPRESS_HTTP_INSTRUMENTATION_KEY is defined.
-from opentelemetry.context import _SUPPRESS_HTTP_INSTRUMENTATION_KEY
+from opentelemetry import trace
 from opentelemetry.instrumentation.urllib3 import URLLib3Instrumentor
-from opentelemetry.instrumentation.utils import _SUPPRESS_INSTRUMENTATION_KEY
+from opentelemetry.instrumentation.utils import (
+    suppress_http_instrumentation,
+    suppress_instrumentation,
+)
 from opentelemetry.propagate import get_global_textmap, set_global_textmap
 from opentelemetry.semconv.trace import SpanAttributes
 from opentelemetry.test.mock_textmap import MockTextMapPropagator
@@ -225,20 +225,17 @@ class TestURLLib3Instrumentor(TestBase):
         URLLib3Instrumentor().instrument()
 
     def test_suppress_instrumentation(self):
-        suppression_keys = (
-            _SUPPRESS_HTTP_INSTRUMENTATION_KEY,
-            _SUPPRESS_INSTRUMENTATION_KEY,
+        suppression_cms = (
+            suppress_instrumentation,
+            suppress_http_instrumentation,
         )
-        for key in suppression_keys:
+        for cm in suppression_cms:
             self.memory_exporter.clear()
 
-            with self.subTest(key=key):
-                token = context.attach(context.set_value(key, True))
-                try:
+            with self.subTest(cm=cm):
+                with cm():
                     response = self.perform_request(self.HTTP_URL)
                     self.assertEqual(b"Hello!", response.data)
-                finally:
-                    context.detach(token)
 
                 self.assert_span(num_spans=0)